refactor(Menu) #420: ddd 구조화 후 인터페이스 다형성 추가
