+++ be/src/main/java/com/foodymoody/be/feed/domain/entity/Menus.java
@@ -1,6 +1,6 @@
 package com.foodymoody.be.feed.domain.entity;
 
-import com.foodymoody.be.menu.domain.Menu;
+import com.foodymoody.be.menu.domain.entity.Menu;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
