+++ be/src/main/java/com/foodymoody/be/feed/infra/persistence/jpa/StoreMoodRepositoryImpl.java
@@ -1,6 +1,5 @@
 package com.foodymoody.be.feed.infra.persistence.jpa;
 
-import com.foodymoody.be.common.exception.MoodNotFoundException;
 import com.foodymoody.be.common.util.ids.StoreMoodId;
 import com.foodymoody.be.feed.domain.entity.StoreMood;
 import com.foodymoody.be.feed.domain.repository.StoreMoodRepository;
