+++ be/src/main/java/com/foodymoody/be/feed/infra/usecase/FeedUseCase.java
@@ -38,8 +38,8 @@ import com.foodymoody.be.image.domain.Image;
 import com.foodymoody.be.member.application.MemberQueryService;
 import com.foodymoody.be.member.application.dto.FeedAuthorSummary;
 import com.foodymoody.be.member.domain.Member;
-import com.foodymoody.be.menu.domain.Menu;
-import com.foodymoody.be.menu.service.MenuService;
+import com.foodymoody.be.menu.application.MenuService;
+import com.foodymoody.be.menu.domain.entity.Menu;
 import com.foodymoody.be.store.application.StoreReadService;
 import java.time.LocalDateTime;
 import java.util.List;
