+++ be/src/main/java/com/foodymoody/be/menu/domain/repository/MenuRepository.java
@@ -1,0 +1,9 @@
+package com.foodymoody.be.menu.domain.repository;
+
+import com.foodymoody.be.menu.domain.entity.Menu;
+
+public interface MenuRepository {
+
+    Menu save(Menu menu);
+
+}
