+++ be/src/test/java/com/foodymoody/be/feed/application/MenuMapperTest.java
@@ -5,8 +5,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import com.foodymoody.be.common.util.ids.IdFactory;
 import com.foodymoody.be.common.util.ids.MenuId;
 import com.foodymoody.be.feed.application.dto.request.FeedRegisterRequestMenu;
-import com.foodymoody.be.menu.domain.Menu;
-import com.foodymoody.be.menu.util.MenuMapper;
+import com.foodymoody.be.menu.domain.entity.Menu;
+import com.foodymoody.be.menu.application.MenuMapper;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
